+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Next Version] - Release Date
+### Fixed
+- login now use the --cluster flag to save the token to config file
+
 ## [0.8.0] - 2017-09-18
 ### Added
 - Command 'deploy list'