+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Next Version] - Release Date
+## [0.9.0] - 2017-10-26
 ### Fixed
 - login now use the --cluster flag to save the token to config file
 - Don't return error on `app info` command if the app doesn't have HPA